Modulo de usuarios, materias y reservas adaptado a 6NF

+++ src/main/java/co/edu/usco/peerlink/repository/MateriaRepository.java
@@ -1,7 +1,4 @@
 package co.edu.usco.peerlink.repository;
-
 import co.edu.usco.peerlink.model.Materia;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MateriaRepository extends JpaRepository<Materia, Long> {
-}
+public interface MateriaRepository extends JpaRepository<Materia, Integer> {}
